requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.4
django-csp==4.0
django-ratelimit==4.1.0
idna==3.11
python-decouple==3.8
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
supervisor==4.3.0
urllib3==2.6.3
wheel==0.46.3
djangorestframework==3.17.1
behave==1.3.3
behave-django==1.9.0
pillow==12.2.0

